+++ src/lib/sparql_queries.py
@@ -2,44 +2,68 @@ from typing import List, TypedDict, Dict
 import streamlit as st
 from lib.sparql_base import query, execute
 from lib.utils import ensure_uri
+from lib.schema import Entity, EntityDetailed, Triple, TripleDetailed
 
 
-##### Model #####
+def insert(triples: List[Triple], graph: str = None) -> None:
+    """
+    From a list of tuples containing the triples in tuple format
+    (('subject', 'predicate', 'object)), insert them in the endpoint,
+    in the given graph.
+    """
 
-class Entity(TypedDict):
-    """Any type of entity."""
+    # Make sure the given graph is a valid URI
+    graph_uri = ensure_uri(graph)
 
-    uri: str
-    label: str
-    comment: str
+    # Since insert can be pretty huge, here we split 
+    # in "smaller insert" of maximum 10k triples.
+    step_nb = 10000
+    for i in range(0, len(triples), step_nb):
+        
+        # Take only selected triples
+        selected_triples = triples[i : i + step_nb]
 
+        # Transform the triples into strings
+        triples_str = '\n'.join(map(lambda triple: triple.to_sparql(), selected_triples))
+        
+        # Prepare the query
+        text = """
+            INSERT DATA {
+                """ + ("GRAPH " + graph_uri + " {" if graph else "") + """
+                """ + triples_str + """
+                """ + ("}" if graph else "") + """
+            }
+        """
 
-class EntityWithClass(TypedDict):
-    """Any type of entity with class (and class label)."""
-
-    uri: str
-    label: str
-    cls: str
-    class_label: str
-
-
-class Triple:
-    """All needed information of a triple."""
-
-    subject: str
-    subject_label: str
-    subject_class: str
-    subject_class_label: str
-    predicate: str
-    predicate_label: str
-    object: str
-    object_label: str
-    object_class: str
-    object_class_label: str
-    isliteral: str
+        # Insert the triples in the endpoint
+        execute(text)
 
 
-##### Functions #####    
+def delete(triples: List[Triple], graph: str = None) -> None:
+    """
+    From a list of tuples containing the triples in ('subject', 'predicate', 'object')
+    format, deleted from the endpoint.
+    triples can be raw triple, or rules with variables.
+    """
+
+    # Make sure the given graph is a valid URI
+    graph_uri = ensure_uri(graph)
+
+    # Transform the triples into strings
+    triples_str = '\n'.join(map(lambda triple: triple.to_sparql(), triples))
+
+    # Prepare query
+    text = """
+        DELETE WHERE {
+            """ + ("GRAPH " + graph_uri + " {" if graph else "") + """
+                """ + triples_str + """
+            """ + ("}" if graph else "") + """
+        }
+    """
+
+    # Execute
+    execute(text)
+ 
 
 @st.cache_data(ttl="1d", show_spinner=False)
 def list_used_classes(graph: str = None) -> List[Entity]:
@@ -49,23 +73,26 @@ def list_used_classes(graph: str = None) -> List[Entity]:
     (ie: objects of 'subject a object' triples).
     """
 
-    # Only if the graph is not default (None)
-    graph_begin = "GRAPH " + ensure_uri(graph) + " {" if graph else ""
-    graph_end = "}" if graph else ""
+    # Make sure the given graph is a valid URI
+    graph_uri = ensure_uri(graph)
 
     # Prepare the query
     text = """
-        SELECT DISTINCT ?uri (COALESCE(?label_, ?uri) as ?label)
+        SELECT DISTINCT 
+            ?uri 
+            (COALESCE(?label_, ?uri) as ?label)
         WHERE { 
-        """ + graph_begin + """
-            ?subject a ?uri .
-            optional { ?uri rdfs:label ?label_ . }
-        """ + graph_end + """
+            """ + ("GRAPH " + graph_uri + " {" if graph else "") + """
+                ?s a ?uri .
+                optional { ?uri rdfs:label ?label_ . }
+            """ + ("}" if graph else "") + """
         }
     """
 
-    # Ensure response is an array
+    # Execute on the endpoint
     response = query(text)
+
+    # Ensure an array is returned
     if not response:
         return []
     return response
@@ -79,30 +106,33 @@ def list_used_properties(graph: str = None) -> List[Entity]:
     (ie: properties of 'subject property object' triples).
     """
 
-    # Only if the graph is not default (None)
-    graph_begin = "GRAPH " + ensure_uri(graph) + " {" if graph else ""
-    graph_end = "}" if graph else ""
+    # Make sure the given graph is a valid URI
+    graph_uri = ensure_uri(graph)
 
     # Prepare the query
     text = """
-        SELECT DISTINCT ?uri (COALESCE(?label_, ?uri) as ?label)
+        SELECT DISTINCT 
+            ?uri 
+            (COALESCE(?label_, ?uri) as ?label)
         WHERE { 
-        """ + graph_begin + """
-            ?subject ?uri ?object .
-            optional { ?uri rdfs:label ?label_ . }
-        """ + graph_end + """
+            """ + ("GRAPH " + graph_uri + " {" if graph else "") + """
+                ?subject ?uri ?object .
+                optional { ?uri rdfs:label ?label_ . }
+            """ + ("}" if graph else "") + """
         }
     """
 
-    # Ensure response is an array
+    # Execute on the endpoint
     response = query(text)
+
+    # Ensure an array is returned
     if not response:
         return []
     return response
 
 
 @st.cache_data(ttl="1d", show_spinner=False)
-def list_entities(label: str = None, cls: str = None, limit: int = None, graph = None) -> List[EntityWithClass]:
+def list_entities(label: str = None, cls: str = None, limit: int = None, graph = None) -> List[EntityDetailed]:
     """
     Fetch the list of entities on the endpoint with:
     Args:
@@ -113,6 +143,8 @@ def list_entities(label: str = None, cls: str = None, limit: int = None, graph =
 
     # Make sure the given class is a valid URI
     class_uri = ensure_uri(cls)
+    # Make sure the given graph is a valid URI
+    graph_uri = ensure_uri(graph)
 
     # Create parts of the request, depending on args:
 
@@ -127,26 +159,28 @@ def list_entities(label: str = None, cls: str = None, limit: int = None, graph =
     # Limit
     limit = f"LIMIT {limit}" if limit else ""
 
-    # Only if the graph is not default (None)
-    graph_begin = "GRAPH " + ensure_uri(graph) + " {" if graph else ""
-    graph_end = "}" if graph else ""
-
     # Build the query
     text = """
-        SELECT DISTINCT ?uri ?label (COALESCE(?cls_, '""" + (cls or 'Unknown') + """') as ?cls) (COALESCE(?class_label_, ?cls) as ?class_label)
+        SELECT DISTINCT 
+            ?uri 
+            ?label 
+            (COALESCE(?cls_, '""" + (cls or 'Unknown') + """') as ?cls) 
+            (COALESCE(?class_label_, ?cls) as ?class_label)
         WHERE {
-            """ + graph_begin + """
-            """ + entity_class + """
-            """ + label_filter + """
-            """ + entity_label + """
-            """ + class_label + """
-            """ + graph_end + """
+            """ + ("GRAPH " + graph_uri + " {" if graph else "") + """
+                """ + entity_class + """
+                """ + label_filter + """
+                """ + entity_label + """
+                """ + class_label + """
+            """ + ("}" if graph else "") + """
         }
         """ + limit + """
     """
 
-    # Ensure response is an array
+    # Execute on the endpoint
     response = query(text)
+
+    # Ensure an array is returned
     if not response:
         return []
     return response
@@ -170,10 +204,18 @@ def list_entity_triples(entity: str, graph=None) -> List[Triple]:
 
     # Prepare the query
     text = """
-        SELECT  ?subject (COALESCE(?subject_label_, 'No label') as ?subject_label) (COALESCE(?subject_class_, 'No class') as ?subject_class) (COALESCE(?subject_class_label_, ?subject_class) as ?subject_class_label)
-                ?predicate (COALESCE(?predicate_label_, ?predicate) as ?predicate_label)
-                ?object (COALESCE(?object_label_, ?object) as ?object_label) (COALESCE(?object_class_, 'No class') as ?object_class) (COALESCE(?object_class_label_, ?object_class) as ?object_class_label)
-                (isLiteral(?object) as ?object_literal)
+        SELECT  
+            ?subject 
+            (COALESCE(?subject_label_, 'No label') as ?subject_label) 
+            (COALESCE(?subject_class_, 'No class') as ?subject_class) 
+            (COALESCE(?subject_class_label_, ?subject_class) as ?subject_class_label)
+            ?predicate 
+            (COALESCE(?predicate_label_, ?predicate) as ?predicate_label)
+            ?object 
+            (COALESCE(?object_label_, ?object) as ?object_label) 
+            (COALESCE(?object_class_, 'No class') as ?object_class) 
+            (COALESCE(?object_class_label_, ?object_class) as ?object_class_label)
+            (isLiteral(?object) as ?object_literal)
         WHERE {
             """ + graph_begin + """
             {
@@ -233,43 +275,17 @@ def list_graphs(_error_location) -> List[Entity]:
     """
 
     # Prepare the query
-    # text = """
-    #     SELECT DISTINCT ?uri ?label ?comment 
-    #     WHERE {
-    #         GRAPH ?uri {
-    #             optional { ?uri rdfs:label ?label . }
-    #             optional { ?uri rdfs:comment ?comment . }
-    #         }
-    #     }
-    # """
-
     text = """
-        SELECT DISTINCT ?uri (CONCAT(COALESCE(?name, ?last_part), ' (', STR(?number), ')') as ?label) ?comment
+        SELECT DISTINCT 
+            ?uri 
+            (COALESCE(?label_, ?uri) as ?label)
+            (COALESCE(?comment_, 'No comment') as ?comment)
         WHERE {
-        {
-            SELECT ?uri ?name (COUNT(*) as ?number) 
-            WHERE {
-                GRAPH ?uri { 
-                    ?s ?p ?o 
-                }
-                OPTIONAL {?uri rdfs:label ?name .}
-                OPTIONAL {?uri rdfs:comment ?comment .}
-            }
-            GROUP BY ?uri  ?name
+            GRAPH ?uri { ?s ?p ?o . }
+            optional { ?uri rdfs:label ?label_ . }
+            optional { ?uri rdfs:comment ?comment_ . }
         }
-        UNION
-        {
-            SELECT (URI('http://my.org/default') as ?uri) (COUNT(*) as ?number) ('DEFAULT' as ?name) ('Default graph' as ?comment)
-            WHERE {
-                { ?s ?p ?o }
-            }
-        }
-        ### noter qu'en Python il faut *2 de antislashes par rapport à SPARQL
-        # en SPARQL \\/ 
-        BIND (replace(str(?uri), "^(.*)(\\\\/)", "") as ?last_part)  
-        }
-    """
-    
+    """    
 
     # Ensure response is an array
     response = query(text, _error_location=_error_location)
@@ -308,96 +324,6 @@ def count_graph_triples(graph: str) -> int:
     if not response:
         return 0
     return int(response[0]['count'])
-
-
-def insert(triples: List[tuple], graph: str = None) -> None:
-    """
-    From a list of tuples containing the triples in tuple format
-    (('subject', 'predicate', 'object)), insert them in the endpoint,
-    in the given graph.
-    """
-
-    # Make sure the given graph is a valid URI
-    graph_uri = ensure_uri(graph)
-
-    # Prepare the query
-    req_insert_begin = "INSERT DATA {"
-    req_insert_end = "}"
-    req_graph_begin = "GRAPH " + graph_uri + " {" if graph else ""
-    req_graph_end = "}" if graph else ""
-
-    # Since insert can be pretty huge, here we split 
-    # in "smaller insert" of maximum 10k triples.
-    step_nb = 10000
-    for i in range(0, len(triples), step_nb):
-        
-        # Take only selected triples
-        selected_triples = triples[i : i + step_nb]
-        
-        # Prepare the query
-        text_begin = f"""
-        {req_insert_begin}
-            {req_graph_begin}
-        """
-        text_end = f"""
-            {req_graph_end}
-        {req_insert_end}
-        """
-
-        # Generate triples
-        texts_triples = []
-        for triple in selected_triples:
-            subject = ensure_uri(triple[0])
-            predicate = ensure_uri(triple[1])
-            object = ensure_uri(triple[2])
-            texts_triples.append(f"{subject} {predicate} {object} .")
-
-        # Build the request text
-        text = text_begin + "\n".join(texts_triples) + text_end
-
-        # Insert the triples in the endpoint
-        execute(text)
-
-
-def delete(triples: List[tuple], graph: str = None) -> None:
-    """
-    From a list of tuples containing the triples in ('subject', 'predicate', 'object')
-    format, deleted from the endpoint.
-    triples can be raw triple, or rules with variables.
-    """
-
-    # Make sure the given graph is a valid URI
-    graph_uri = ensure_uri(graph)
-
-    # Prepare the query
-    req_insert_begin = "DELETE WHERE {"
-    req_insert_end = "}"
-    req_graph_begin = "GRAPH " + graph_uri + " {" if graph else ""
-    req_graph_end = "}" if graph else ""
-
-    # Prepare the request
-    text_begin = f"""
-    {req_insert_begin}
-        {req_graph_begin}
-    """
-    text_end = f"""
-        {req_graph_end}
-    {req_insert_end}
-    """
-
-    # Generate all triples
-    texts_triples = []
-    for triple in triples:
-        subject = ensure_uri(triple[0])
-        predicate = ensure_uri(triple[1])
-        object = ensure_uri(triple[2])
-        texts_triples.append(f"{subject} {predicate} {object} .")
-
-    # Build the request text
-    text = text_begin + "\n".join(texts_triples) + text_end
-
-    # Execute
-    execute(text)
 
 
 def get_objects_of(subject: str, property: str = None):